Improve groupMapReduce performance
Reduce the amount of lookups performed inside a groupMapReduce by roughly half.

Every iteration inside of `groupMapReduce` would previously hit both `get()` and `put()`,
both of which result in lookup operations on the `Map`.

This patch eliminates all the additional `put()`s after the first `put()` for a given key by using
a wrapper `Buffer` for the value stored so that the `Buffer` can be updated directly after the
`get()` and `reduce` operation are performed.

NOTE:
I also did try to use `updateWith()` rather than using a `Buffer` wrapper around the value, but
it ended up being even slower than the original version of `groupMapReduce()`.